requirements.txt
alabaster==0.7.13
asgiref==3.8.1
babel==2.16.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.9
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
docutils==0.20.1
idna==3.10
imagesize==1.4.1
importlib-metadata==8.5.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
pillow==10.4.0
pockets==0.9.1
pprintpp==0.4.0
psycopg2-binary==2.9.10
pygments==2.18.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.1.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django==2.5
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-napoleon==0.7
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
